requirements.txt
streamlit>=1.35.0
plotly>=5.18.0
numpy>=1.24.0
scipy>=1.11.0
yfinance>=0.2.31
anthropic>=0.28.0
finnhub-python
python-dotenv
pandas>=2.0.0
